Change packet lengths to sized types
IPv4 packet header sizes fit in a u8 value, and all IP/Ethernet-related
packet total sizes fit in a u16 value. This changes those sizes to use
smaller types that don't ever need to be converted to smaller types.

+++ src/vmm/src/dumbo/pdu/mod.rs
@@ -79,28 +79,26 @@ fn compute_checksum<T: NetworkBytes + Debug>(
     dst_addr: Ipv4Addr,
     protocol: ChecksumProto,
 ) -> u16 {
-    // TODO: Is u32 enough to prevent overflow for the code in this function? I think so, but it
-    // would be nice to double-check.
-    let mut sum = 0u32;
+    let mut sum = 0usize;
 
-    let a = u32::from(src_addr);
+    let a = u32::from(src_addr) as usize;
     sum += a & 0xffff;
     sum += a >> 16;
 
-    let b = u32::from(dst_addr);
+    let b = u32::from(dst_addr) as usize;
     sum += b & 0xffff;
     sum += b >> 16;
 
     let len = bytes.len();
-    sum += protocol as u32;
-    sum += len as u32;
+    sum += protocol as usize;
+    sum += len;
 
     for i in 0..len / 2 {
-        sum += u32::from(bytes.ntohs_unchecked(i * 2));
+        sum += usize::from(bytes.ntohs_unchecked(i * 2));
     }
 
     if len % 2 != 0 {
-        sum += u32::from(bytes[len - 1]) << 8;
+        sum += usize::from(bytes[len - 1]) << 8;
     }
 
     while sum >> 16 != 0 {
